--- apps/audits/signals.py
@@ -1,6 +1,0 @@
-from django.dispatch import Signal
-
-
-post_activity_log = Signal(
-    providing_args=('resource_id', 'detail', 'detail_url')
-)
